Create child directory and move myDragons
